Added methods: Operator::add_delay_signal_no_reset                Operator::add_delay_signal_bus_no_reset
* the "_bus" in the name of the second method determines all the signals to be represented as 
vectors, even signals having 1 bit width. When this is used for example:

	add_delay_signal_bus_no_reset("a",1);

will create the corresponding code in the .vhdl

	... a: std_logic_vector(0 downto 0);

This option is useful in homogeneously manipulation of vectors in the case of pipelining 
operations. 

* the "_no_reset" means that registers added will not have reset. This simple option reduces  the number of 
slices used ( I saw that when generating registers with reset, Xilinx was informing me that if I remove the reset I 
save something in terms of area).